Remove custom HTTP Client pool
